+++ test/pages/journal/infinite_journal_page_test.dart
@@ -36,8 +36,7 @@ void main() {
   var mockPersistenceLogic = MockPersistenceLogic();
   final mockEntitiesCacheService = MockEntitiesCacheService();
 
-  final entryTypeStrings =
-      entryTypes.map((e) => e.typeName).whereType<String>().toList();
+  final entryTypeStrings = entryTypes.toList();
 
   group('JournalPage Widget Tests - ', () {
     setUpAll(() {
@@ -234,6 +233,47 @@ void main() {
         (tester.firstWidget(find.byIcon(MdiIcons.star)) as Icon).color,
         darkTheme.starredGold,
       );
+    });
+
+    testWidgets('page shows task filter', (tester) async {
+      when(
+        () => mockJournalDb.watchTaskCount(any()),
+      ).thenAnswer(
+        (_) => Stream<int>.fromIterable([10]),
+      );
+
+      when(
+        () => mockJournalDb.watchJournalEntities(
+          types: entryTypeStrings,
+          starredStatuses: [true, false],
+          privateStatuses: [true, false],
+          flaggedStatuses: [1, 0],
+          ids: null,
+          limit: 50,
+        ),
+      ).thenAnswer(
+        (_) => Stream<List<JournalEntity>>.fromIterable([
+          [testTask]
+        ]),
+      );
+
+      await tester.pumpWidget(
+        makeTestableWidgetWithScaffold(
+          BlocProvider<AudioPlayerCubit>(
+            create: (BuildContext context) => AudioPlayerCubit(),
+            lazy: false,
+            child: const InfiniteJournalPage(),
+          ),
+        ),
+      );
+
+      await tester.pumpAndSettle();
+
+      final tasksSegment = find.text('Tasks');
+      expect(tasksSegment, findsOneWidget);
+
+      await tester.tap(tasksSegment);
+      await tester.pumpAndSettle();
     });
 
     testWidgets('page is rendered with weight entry', (tester) async {