+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed:
 - Journal card styling
+- Entry type filter styling
+- Task filter styling
 
 ## [0.8.233] - 2023-01-13
 ### Changed: